% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bioanth.R
\docType{package}
\name{bioanth}
\alias{bioanth}
\alias{bioanth-package}
\title{bioanth: Datasets useful for biological and forensic anthropology.}
\description{
\code{bioanth} provides datasets useful for biological and forensic
anthropology. In addition, it includes helper functions for working with
dataframes.
}
\details{
This package contains the following datasets:
\itemize{
  \item \code{howell}: Howells Craniometric Data Set
  \item \code{howelltest}: Howells Craniometric Test Data Set
  \item \code{goldman}: Goldman Osteometric Data Set
}

and the following functions:
\itemize{
  \item \code{lower_names}: Converts column names of a data frame to
    lower (or upper) case.
  \item \code{lower_factors}: Converts factor labels to
    lower (or upper) case.
}
}

