% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goldman}
\alias{goldman}
\title{Goldman Osteometric Data Set}
\format{A data frame with 1538 rows and 66 variables:
\itemize{
  \item \code{Inst}: Collection housing remains
  \item \code{ID}: Unique case ID
  \item \code{Sex}: Male or female, determined from os coxae
  \item \code{Age}: Age range based on pubic symphysis and auricular surface
  \item \code{NOTE}: For most skeletons, site of origin location
  \item \code{Location}: Modern country of origin or
    state (when United States)
  \item \code{LHUM}: Left humerus: 0 = present, 1 = absent
  \item \code{RHUM}: Right humerus: 0 = present, 1 = absent
  \item \code{LRAD}: Left radius: 0 = present, 1 = absent
  \item \code{RRAD}: Right radius: 0 = present, 1 = absent
  \item \code{LFEM}: Left femur: 0 = present, 1 = absent
  \item \code{RFEM}: Right femur: 0 = present, 1 = absent
  \item \code{LTIB}: Left tibia: 0 = present, 1 = absent
  \item \code{RTIB}: Right tibia: 0 = present, 1 = absent
  \item \code{OSCX}: Os coxae: 0 = present, 1 = absent
  \item \code{LHML}: Left humerus maximum length, mm
  \item \code{LHEB}: Left humerus epicondylar breadth, mm
  \item \code{LHHD}: Left humerus head diameter, mm
  \item \code{LHMLD}: Left humerus 50\% diaphyseal mediolateral
    diameter, mm
  \item \code{LHAPD}: Left humerus 50\% Diaphyseal anteroposterior
    diameter, mm
  \item \code{RHML}: Right humerus maximum length, mm
  \item \code{RHEB}: Right humerus epicondylar breadth, mm
  \item \code{RHHD}: Right humerus head diameter, mm
  \item \code{RHMLD}: Right humerus 50\% diaphyseal mediolateral
    diameter, mm
  \item \code{RHAPD}: Right humerus 50\% Diaphyseal anteroposterior
    diameter, mm
  \item \code{LRML}: Left radius maximum length, mm
  \item \code{LRMLD}: Left radius 50\% diaphyseal mediolateral diameter, mm
  \item \code{LRAPD}: Left radius 50\% diaphyseal anteroposterior
    diameter, mm
  \item \code{RRML}: Right radius maximum length, mm
  \item \code{RRMLD}: Right radius 50\% diaphyseal mediolateral diameter, mm
  \item \code{RRAPD}: Right radius 50\% diaphyseal anteroposterior
    diameter, mm
  \item \code{LFML}: Left femur maximum length, mm
  \item \code{LFBL}: Left femur femur bicondylar length, mm
  \item \code{LFEB}: Left femur epicondylar mediolateral breadth, mm
  \item \code{LFAB}: Left femur distal articular mediolateral breadth, mm
  \item \code{LFHD}: Left femur head anteroposterior diameter, mm
  \item \code{LFMLD}: Left femur 50\% Diaphyseal Mediolateral Diameter, mm
  \item \code{LFAPD}: Left femur 50\% diaphyseal anteroposterior diameter, mm
  \item \code{RFML}: Right femur maximum length, mm
  \item \code{RFBL}: Right femur bicondylar length, mm
  \item \code{RFEB}: Right femur epicondylar mediolateral breadth, mm
  \item \code{RFAB}: Right femur distal articular mediolateral breadth, mm
  \item \code{RFHD}: Right femur head anteroposterior diameter, mm
  \item \code{RFMLD}: Right femur 50\% Diaphyseal Mediolateral Diameter, mm
  \item \code{RFAPD}: Right femur 50\% diaphyseal anteroposterior
    diameter, mm
  \item \code{LTML}: Left tibia maximum length, mm
  \item \code{LTPB}: left tibia plateau mediolateral (bicondylar) breadth, mm
  \item \code{LTMLD}: Left tibia 50\% diaphyseal mediolateral diameter, mm
  \item \code{LTAPD}: Left tibia 50\% diaphyseal anteroposterior diameter, mm
  \item \code{RTML}: Right tibia maximum length, mm
  \item \code{RTPB}: Right tibia plateau mediolateral
    (bicondylar) breadth, mm
  \item \code{RTMLD}: Right tibia 50\% diaphyseal mediolateral diameter, mm
  \item \code{RTAPD}: Right tibia 50\% diaphyseal anteroposterior
    diameter, mm
  \item \code{BIB}: Bi-iliac breadth, mm
  \item \code{LIBL}: Left Maximum iliac blade length, mm
  \item \code{RIBL}: Right Maximum iliac blade length, mm
  \item \code{LAcH}: Left maximum acetabular height, mm
  \item \code{RAcH}: Right maximum acetabular height, mm
  \item \code{Brachial}
  \item \code{Crural}
  \item \code{IL.UL.LL}
  \item \code{IL.LL.UL}
  \item \code{CBR.FHD}
  \item \code{McH.FHD}
  \item \code{GRINE.FHD}
  \item \code{AVG.FHD}
}}
\source{
\url{http://web.utk.edu/~auerbach/GOLD.htm}
}
\usage{
goldman
}
\description{
The Goldman Osteometric Data Set consists of postcranial skeletal metric
measurements, sex estimates, and broad age estimates made by
Dr. Benjamin Auerbach between 2001 and 2003. These represent a global sample
of humans dating throughout the Holocene, though most samples are from
preindustrial groups. A total of over 1500 individuals from 59 locations are
included in the data set. Only measurements from the upper limb (humerus &
radius), lower limb (femur & tibia), and pelvis were taken.
}
\section{Citation}{

\itemize{
  \item Auerbach BM, & Ruff CB. 2004. Human body mass estimation:
    a comparison of "morphometric" and "mechanical" methods.
    American Journal of Physical Anthropology 125:331-342.

  \item Auerbach BM, & Ruff CB. 2006. Limb bone bilateral asymmetry:
    variability and commonality among modern humans.
    Journal of Human Evolution 50:203-218.
}
}
\keyword{datasets}

