% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{howell}
\alias{howell}
\title{Howells Craniometric Data Set}
\format{A data frame with 2524 rows and 89 variables:}
\source{
\url{http://web.utk.edu/~auerbach/HOWL.htm}
}
\usage{
howell
}
\description{
The William W. Howells Craniometric Data Set consists of cranial metric
measurements and sex estimates made by Dr. William Howells between 1965 and
1980. These represent a global sample of human groups considered by
Dr. Howells to represent populations, all of which date to the later
Holocene. Over 2500 individuals from 28 populations were measured by
Dr. Howells. Craniometric dimensions in the data set are linear measurements
taken with calipers, and include lengths, chords, subtenses, angles,
and radii.
}
\section{Variables}{

\itemize{
  \item \code{ID}: Unique case ID
  \item \code{Sex}: M or F
  \item \code{PopNum}: Population number
  \item \code{Population}: Population name
  \item \code{GOL}: mm
  \item \code{NOL}: mm
  \item \code{BNL}: mm
  \item \code{BBH}: mm
  \item \code{XCB}: mm
  \item \code{XFB}: mm
  \item \code{ZYB}: mm
  \item \code{AUB}: mm
  \item \code{WCB}: mm
  \item \code{ASB}: mm
  \item \code{BPL}: mm
  \item \code{NPH}: mm
  \item \code{NLH}: mm
  \item \code{JUB}: mm
  \item \code{NLB}: mm
  \item \code{MAB}: mm
  \item \code{MDH}: mm
  \item \code{MDB}: mm
  \item \code{OBH}: mm
  \item \code{OBB}: mm
  \item \code{DKB}: mm
  \item \code{NDS}: mm
  \item \code{WNB}: mm
  \item \code{SIS}: mm
  \item \code{ZMB}: mm
  \item \code{SSS}: mm
  \item \code{FMB}: mm
  \item \code{NAS}: mm
  \item \code{EKB}: mm
  \item \code{DKS}: mm
  \item \code{IML}: mm
  \item \code{XML}: mm
  \item \code{MLS}: mm
  \item \code{WMH}: mm
  \item \code{SOS}: mm
  \item \code{GLS}: mm
  \item \code{STB}: mm
  \item \code{FRC}: mm
  \item \code{FRS}: mm
  \item \code{FRF}: mm
  \item \code{PAC}: mm
  \item \code{PAS}: mm
  \item \code{PAF}: mm
  \item \code{OCC}: mm
  \item \code{OCS}: mm
  \item \code{OCF}: mm
  \item \code{FOL}: mm
  \item \code{NAR}: mm
  \item \code{SSR}: mm
  \item \code{PRR}: mm
  \item \code{DKR}: mm
  \item \code{ZOR}: mm
  \item \code{FMR}: mm
  \item \code{EKR}: mm
  \item \code{ZMR}: mm
  \item \code{AVR}: mm
  \item \code{BRR}: mm
  \item \code{VRR}: mm
  \item \code{LAR}: mm
  \item \code{OSR}: mm
  \item \code{BAR}: mm
  \item \code{NAA}: mm
  \item \code{PRA}: mm
  \item \code{BAA}: mm
  \item \code{NBA}: mm
  \item \code{BBA}: mm
  \item \code{BRA}: mm
  \item \code{SSA}: mm
  \item \code{NFA}: mm
  \item \code{DKA}: mm
  \item \code{NDA}: mm
  \item \code{SIA}: mm
  \item \code{FRA}: mm
  \item \code{PAA}: mm
  \item \code{OCA}: mm
  \item \code{RFA}: mm
  \item \code{RPA}: mm
  \item \code{ROA}: mm
  \item \code{BSA}: mm
  \item \code{SBA}: mm
  \item \code{SLA}: mm
  \item \code{TBA}: mm
}
}

\section{Citation}{

Howells WW. 1973. Cranial Variation in Man. A Study by Multivariate Analysis
of Patterns of Differences Among Recent Human Populations.
Papers of the Peabody Museum of Archeology and Ethnology, vol. 67, pp. 259.
Cambridge, Mass.: Peabody Museum.

Howells WW. 1989. Skull Shapes and the Map. Craniometric Analyses in the
Dispersion of Modern Homo.
Papers of the Peabody Museum of Archaeology and Ethnology, vol. 79, pp. 189.
Cambridge, Mass.: Peabody Museum.

Howells WW. 1995. Who’s Who in Skulls. Ethnic Identification of Crania from
Measurements.
Papers of the Peabody Museum of Archaeology and Ethnology, vol. 82, pp. 108.
Cambridge, Mass.: Peabody Museum.
}
\seealso{
Other howells: \code{\link{howelltest}}
}
\keyword{datasets}

