% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{howelltest}
\alias{howelltest}
\title{Howell Craniometric Test Data Set}
\format{A data frame with 524 rows and 89 variables:}
\source{
\url{http://web.utk.edu/~auerbach/HOWL.htm}
}
\usage{
howelltest
}
\description{
The William W. Howells Craniometric Data Set consists of craniometric
measurements taken from 2524 human crania from 28 populations.
Up to 82 measurements were obtained from these crania. In addition,
Dr. Howells obtained data from 524 "test" crania (crania not included in the
main data set).
}
\section{Variables}{

\itemize{
  \item \code{ID}: Unique case ID
  \item \code{Sex}: M or F
  \item \code{Tribe}: Tribe name
  \item \code{Origin}: Origin location
  \item \code{Museum.location}: Museum location
  \item \code{Museum.number}: Museum number
  \item \code{Comments}: Notes
  \item \code{GOL}: mm
  \item \code{NOL}: mm
  \item \code{BNL}: mm
  \item \code{BBH}: mm
  \item \code{XCB}: mm
  \item \code{XFB}: mm
  \item \code{ZYB}: mm
  \item \code{AUB}: mm
  \item \code{WCB}: mm
  \item \code{ASB}: mm
  \item \code{BPL}: mm
  \item \code{NPH}: mm
  \item \code{NLH}: mm
  \item \code{JUB}: mm
  \item \code{NLB}: mm
  \item \code{MAB}: mm
  \item \code{MDH}: mm
  \item \code{MDB}: mm
  \item \code{OBH}: mm
  \item \code{OBB}: mm
  \item \code{DKB}: mm
  \item \code{NDS}: mm
  \item \code{WNB}: mm
  \item \code{SIS}: mm
  \item \code{ZMB}: mm
  \item \code{SSS}: mm
  \item \code{FMB}: mm
  \item \code{NAS}: mm
  \item \code{EKB}: mm
  \item \code{DKS}: mm
  \item \code{IML}: mm
  \item \code{XML}: mm
  \item \code{MLS}: mm
  \item \code{WMH}: mm
  \item \code{SOS}: mm
  \item \code{GLS}: mm
  \item \code{STB}: mm
  \item \code{FRC}: mm
  \item \code{FRS}: mm
  \item \code{FRF}: mm
  \item \code{PAC}: mm
  \item \code{PAS}: mm
  \item \code{PAF}: mm
  \item \code{OCC}: mm
  \item \code{OCS}: mm
  \item \code{OCF}: mm
  \item \code{FOL}: mm
  \item \code{NAR}: mm
  \item \code{SSR}: mm
  \item \code{PRR}: mm
  \item \code{DKR}: mm
  \item \code{ZOR}: mm
  \item \code{FMR}: mm
  \item \code{EKR}: mm
  \item \code{ZMR}: mm
  \item \code{AVR}: mm
  \item \code{BRR}: mm
  \item \code{VRR}: mm
  \item \code{LAR}: mm
  \item \code{OSR}: mm
  \item \code{BAR}: mm
  \item \code{NAA}: mm
  \item \code{PRA}: mm
  \item \code{BAA}: mm
  \item \code{NBA}: mm
  \item \code{BBA}: mm
  \item \code{BRA}: mm
  \item \code{SSA}: mm
  \item \code{NFA}: mm
  \item \code{DKA}: mm
  \item \code{NDA}: mm
  \item \code{SIA}: mm
  \item \code{FRA}: mm
  \item \code{PAA}: mm
  \item \code{OCA}: mm
  \item \code{RFA}: mm
  \item \code{RPA}: mm
  \item \code{ROA}: mm
  \item \code{BSA}: mm
  \item \code{SBA}: mm
  \item \code{SLA}: mm
  \item \code{TBA}: mm
}
}

\section{Citation}{

\itemize{
  \item Howells WW. 1973. Cranial Variation in Man. A Study by Multivariate
    Analysis of Patterns of Differences Among Recent Human Populations.
    Papers of the Peabody Museum of Archeology and Ethnology, vol. 67,
    pp. 259.
    Cambridge, Mass.: Peabody Museum.

  \item Howells WW. 1989. Skull Shapes and the Map. Craniometric Analyses
    in the Dispersion of Modern Homo.
    Papers of the Peabody Museum of Archaeology and Ethnology, vol. 79,
    pp. 189.
    Cambridge, Mass.: Peabody Museum.

  \item Howells WW. 1995. Who's Who in Skulls. Ethnic Identification of
    Crania from Measurements.
    Papers of the Peabody Museum of Archaeology and Ethnology, vol. 82,
    pp. 108.
    Cambridge, Mass.: Peabody Museum.
}
}
\seealso{
Other howells: \code{\link{howell}}
}
\keyword{datasets}

