% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cases.R
\name{lower_factors}
\alias{lower_factors}
\title{Factor labels to lower (or upper) case.}
\usage{
lower_factors(.data, ..., case = "lower")
}
\arguments{
\item{.data}{A data frame.}

\item{...}{Comma separated list of unquoted expressions. Defaults to all
columns of class factor.}

\item{case}{A character vector \code{"upper"} or \code{"lower"} (default).
Note: case must be a named argument if specified
(e.g. \code{case = "upper"}).}
}
\value{
A data frame.
}
\description{
\code{lower_factors} acts on columns of a data frame and converts factor
labels to lower case (by default) or upper case (with case = "upper").
\code{lower_factors} plays nice with \code{\%>\%} from dplyr/magrittr.
}
\examples{
lower_factors(iris, Species, case = "upper")

# Lower case is default
lower_factors(iris)

# Standard Evaluation with lower_factors_
lower_factors_(iris, ~Species, case = "upper")
}
\seealso{
Other case.functions: \code{\link{lower_names}}
}

