% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cases.R
\name{lower_names}
\alias{lower_names}
\title{Convert column names to lower (or upper) case.}
\usage{
lower_names(.data, ..., case = "lower")
}
\arguments{
\item{.data}{A data frame.}

\item{...}{Comma separated list of unquoted expressions. Defaults to all
columns.}

\item{case}{A character vector \code{"upper"} or \code{"lower"} (default).
Note: case must be a named argument if specified
(e.g. \code{case = "upper"}).}
}
\value{
A data frame.
}
\description{
\code{lower_names} acts on the column names of a data frame and converts
them to lower case (by default) or upper case (with case = "upper").
\code{lower_names} plays nice with \code{\%>\%} from dplyr/magrittr.
}
\examples{
lower_names(iris, case = "upper")

# Lower case is default
lower_factors(iris)

#Specific columns
lower_names(iris, Sepal.Length, Species)

# Standard Evaluation with lower_names_
lower_factors_(iris, ~Species, case = "upper")
}
\seealso{
Other case.functions: \code{\link{lower_factors}}
}

