% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cases.R
\name{lower_names_}
\alias{lower_names_}
\title{Column names to lower (or upper) case Standard Evaluation.}
\usage{
lower_names_(.data, ..., .dots, case = "lower")
}
\arguments{
\item{.data}{A data frame.}

\item{...}{Comma separated list of unquoted expressions. Defaults to all
columns.}

\item{.dots}{lazy_dots from \code{lower_names}.}

\item{case}{A character vector \code{"upper"} or \code{"lower"} (default).
Note: case must be a named argument if specified
(e.g. \code{case = "upper"}).}
}
\value{
A data frame.
}
\description{
\code{lower_names_} is the standard evaluation version of \code{lower_names}.
Dots arguments should be quoted with ~. This function is mostly for
use in programming.
}
\examples{
lower_names_(iris, ~Sepal.Length, ~Species)
}
\seealso{
Other case functions: \code{\link{lower_factors_}};
  \code{\link{lower_factors}}; \code{\link{lower_names}}
}

