% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bioanth.R
\docType{package}
\name{bioanth}
\alias{bioanth}
\alias{bioanth-package}
\title{bioanth: Datasets useful for biological and forensic anthropology.}
\description{
\code{bioanth} provides datasets useful for biological and forensic
anthropology. In addition, it includes helper functions for working with
dataframes.
}
\details{
This package contains the following datasets:
\itemize{
  \item Howell's craniometrics
  \item Goldman
}
}

