% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cases.R
\name{lower_factors_}
\alias{lower_factors_}
\title{Factor labels to lower (or upper) case Standard Evaluation.}
\usage{
lower_factors_(.data, ..., .dots, case = "lower")
}
\arguments{
\item{.data}{A data frame.}

\item{...}{Comma separated list of unquoted expressions. Defaults to all
columns of class factor.}

\item{.dots}{lazy_dots object from \code{lower_factors}.}

\item{case}{A character vector \code{"upper"} or \code{"lower"} (default).
Note: case must be a named argument if specified
(e.g. \code{case = "upper"}).}
}
\value{
A data frame.
}
\description{
\code{lower_factors_} is the standard evaluation version of
\code{lower_factors}. Dots arguments should be quoted with ~. This function
is mostly for use in programming.
}
\examples{
lower_factors_(iris, ~Species, case = "upper")
}
\seealso{
Other case functions: \code{\link{lower_factors}};
  \code{\link{lower_names_}}; \code{\link{lower_names}}
}

